% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.mvtapered.R
\name{ergm.mvtapered}
\alias{ergm.mvtapered}
\title{Fits a Mean-Valued Tapered ERGM}
\usage{
ergm.mvtapered(formula, r = 2, mv = NULL, tapering.centers = NULL,
  family = "taper", taper.terms = "dependent",
  control = control.ergm(), ...)
}
\arguments{
\item{formula}{An ergm formula to fit}

\item{r}{The scaling factor to use for the hueristic of setting beta equal to r the standard deviation of the observed statistics}

\item{tapering.centers}{The centers of the tapering terms. If null, these are taken to be the mean value parameters.}

\item{control}{An object of class control.ergm. Passed to the ergm function.}

\item{...}{Additional arguments to ergm.}

\item{beta}{The tapering parameters. If not null, these override the hueristics.}
}
\description{
Fits a Mean-Valued Tapered ERGM
}
\examples{
\dontrun{
data(sampson)
fit <- ergm.tapered(samplike ~ edges + triangles())
summary(fit)
}
}
\references{
Fellows, I and Handcock, MS (2017). Removing Phase Transitions from Gibbs Measures. Proceedings of Machine Learning Research, 54:289-297.
}
