% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.tapered-internal.R
\name{ergm.tapered-internal}
\alias{ergm.tapered-internal}
\alias{llik.fun.Kpenalty}
\alias{llik.grad.Kpenalty.numDeriv}
\alias{llik.hessian.Kpenalty.numDeriv}
\alias{llik.fun.obs.Kpenalty}
\alias{llik.grad.obs.Kpenalty.numDeriv}
\alias{llik.hessian.obs.Kpenalty.numDeriv}
\title{Internal ergm.tapered Objects}
\description{
Internal \code{\link[=ergm.tapered-package]{ergm}} functions.
}
\details{
Most of these are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}
\seealso{
ergm.tapered-package
}
\keyword{internal}
