% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.tapered.R
\name{ergm.tapered}
\alias{ergm.tapered}
\title{Fits a Tapered ERGM}
\usage{
ergm.tapered(formula, r = 2, beta = NULL, tau = NULL,
  tapering.centers = NULL, target.stats = NULL, family = "taper",
  taper.terms = "dependent", control = control.ergm(MCMLE.termination =
  "confidence"), ...)
}
\arguments{
\item{formula}{An ergm formula to fit}

\item{r}{The scaling factor to use for the hueristic of setting beta equal to r standard deviations of the observed statistics}

\item{beta}{The tapering parameters, expressed as in Fellows and Handcock (2017). If not NULL, these override the hueristics (r).}

\item{tau}{The tapering parameters, expressed as natural parameters. If not NULL, these override the beta and the hueristics (r).}

\item{tapering.centers}{The centers of the tapering terms. If null, these are taken to be the mean value parameters.}

\item{control}{An object of class control.ergm. Passed to the ergm function.}

\item{...}{Additional arguments to ergm.}
}
\description{
Fits a Tapered ERGM
}
\examples{
\dontrun{
data(sampson)
fit <- ergm.tapered(samplike ~ edges + triangles())
summary(fit)
}
}
\references{
Fellows, I and Handcock, MS (2017). Removing Phase Transitions from Gibbs Measures. Proceedings of Machine Learning Research, 54:289-297.
}
