% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.M4.R
\name{InitErgmTerm.M4}
\alias{InitErgmTerm.M4}
\title{M4}
\usage{
# binary: M4(formula=NULL, coef=NULL, m=NULL)
}
\arguments{
\item{formula}{a valid formula for a  standard ERGM}

\item{coef}{a numeric vector of coefficients giving the penalty coefficients \eqn{\beta} for the tapering of the terms.
If \code{NULL} is passed, the tapering coefficients are set to \code{1/(4*summary(formula))}, the default
in Fellows and Handcock (2017).
If a numeric vector is given, there are interpreted as the tapering coefficients of the terms in the
model, including the terms enclosed in \code{offset()}.
If a numeric scalar is given, it is interpreted as a multiplier of the default tapering coefficients , that is,}

\item{m}{numeric vector. If given, is
the value of the network statistic relative to which the model is
tapered. If omitted, it defaults to that of the model's LHS
network if \code{formula} is one-sided and that of the network on
the LHS of \code{formula} if it is two-sided.}
}
\description{
Adds the fourth moment of each statistics specified by the terms in \code{formula} to the model.
This is sometimes applied directly to control the fourth moment of the standard
statistics, but more typically as part of a Tapered ERGM.
}
\concept{tapered}
