% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmTerm.tapered.R
\name{InitWtErgmTerm.Taper}
\alias{InitWtErgmTerm.Taper}
\title{Weighted Taper}
\usage{
# binary: Taper(formula=NULL, coef=NULL, m=NULL)
}
\arguments{
\item{formula}{a valid formula for a  standard ERGM}

\item{coef}{a numeric vector of coefficients giving the penalty coefficients \eqn{\beta} for the tapering of the terms.
If \code{NULL} is passed, the tapering coefficients are set to \code{1/(4*summary(formula))}, the default
in Fellows and Handcock (2017).
If a numeric vector is given, there are interpreted as the tapering coefficients of the terms in the
model, including the terms enclosed in \code{offset()}.
If a numeric scalar is given, it is interpreted as a multiplier of the default tapering coefficients , that is,}

\item{m}{numeric vector. If given, is
the value of the network statistic relative to which the model is
tapered. If omitted, it defaults to that of the model's LHS
network if \code{formula} is one-sided and that of the network on
the LHS of \code{formula} if it is two-sided.}
}
\description{
Adds the terms specified in \code{formula} to the model \emph{and}
imposes the quadratic penalty of Fellows and Handcock
(2017).
}
\concept{tapered}
