% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ergm.tapered.R
\name{simulate_ergm.tapered}
\alias{simulate_ergm.tapered}
\title{Simulates from a Tapered ERGM formula}
\usage{
simulate_ergm.tapered(
  formula,
  coef,
  nsim = 1,
  r = 2,
  beta = NULL,
  tau = NULL,
  tapering.centers = NULL,
  family = "taper",
  taper.terms = "all",
  response = NULL,
  constraints = ~.,
  reference = ~Bernoulli,
  seed = NULL,
  control = control.simulate.formula(),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{formula}{An ergm formula to fit}

\item{coef}{Vector of parameter values for the model from which the
sample is to be drawn.}

\item{nsim}{Number of networks to be randomly drawn from the given
distribution on the set of all networks, returned by the Metropolis-Hastings
algorithm.}

\item{r}{The scaling factor to use for the heuristic of setting beta equal to r standard deviations of the observed statistics}

\item{beta}{The tapering parameters, expressed as in Fellows and Handcock (2017). If not NULL, these override the heuristics (r).}

\item{tau}{The tapering parameters, expressed as natural parameters. If not NULL, these override the beta and the heuristics (r).}

\item{family}{The type of tapering used. This should either be the \code{stereo} or \code{taper}, the
tapering model of Fellows and Handcock (2017).}

\item{taper.terms}{Specification of the tapering used. If the character variable "dependence" then all the dependent
terms are tapered. If the character variable "all" then all terms are tapered.
It can also be the RHS of a formula giving the terms to be tapered.}

\item{response}{{Name of the edge attribute whose value is to be
modeled in the valued ERGM framework. Defaults to \code{NULL} for
simple presence or absence, modeled via a binary ERGM.}}

\item{constraints}{{A formula specifying one or more constraints
on the support of the distribution of the networks being modeled,
using syntax similar to the \code{formula} argument, on the
right-hand side. See \link[=ergm-constraints]{ERGM constraints}
\code{\link{ergm}} for details.}}

\item{reference}{{A one-sided formula specifying
the reference measure (\eqn{h(y)}) to be used.
See help for \link[=ergm-references]{ERGM reference measures} implemented in the
\strong{\link[=ergm-package]{ergm}} package.}}

\item{seed}{the random seed to use; see \code{\link[=simulate]{simulate()}}.}

\item{control}{An object of class control.ergm. Passed to the ergm function.}

\item{verbose}{A \code{logical}: if this is
\code{TRUE}, the program will print out additional
information about the progress of estimation.}

\item{...}{Additional arguments to \code{\link{ergm}}.}
}
\value{
An object of class c('tapered.ergm','ergm') containing the fit model. In addition to all of the ergm items,
this object contains tapering.centers, tapering.coef and orig.formula. tapering.centers are the centers for the tapering term.
tapering.coef are the tapering coefficients = 1/ beta^2. orig.formula is the formula passed into ergm.tapered.
}
\description{
Simulates from a Tapered ERGM formula
}
\examples{
\dontrun{
data(sampson)
fit <- ergm.tapered(samplike ~ edges + triangles())
summary(fit)
}
}
\references{
\itemize{
\itemize{
\item Fellows, I. and M. S. Handcock (2017),
Removing Phase Transitions from Gibbs Measures. Volume 54 of
Proceedings of Machine Learning Research, Fort Lauderdale,
FL, USA, pp. 289–297. PMLR.
\item Blackburn, B. and M. S. Handcock (2022),
Practical Network Modeling via Tapered Exponential-family Random Graph Models.
Journal of Computational and Graphical Statistics
\doi{10.1080/10618600.2022.2116444}.
}

}
}
