% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ergm.tapered.R
\name{summary.ergm.tapered}
\alias{summary.ergm.tapered}
\alias{print.summary.ergm.tapered}
\title{Summarizing Tapered and Kurtosis ERGM Model Fits}
\usage{
\method{summary}{ergm.tapered}(
  object,
  ...,
  correlation = FALSE,
  covariance = FALSE,
  total.variation = TRUE
)
}
\arguments{
\item{object}{an object of class \code{"ergm.tapered"}, usually, a result
of a call to \code{\link{ergm.tapered}}.}

\item{\dots}{Arguments to \code{\link{logLik.ergm}}}

\item{correlation}{logical; if \code{TRUE}, the correlation matrix
of the estimated parameters is returned and printed.}

\item{covariance}{logical; if \code{TRUE}, the covariance matrix of
the estimated parameters is returned and printed.}

\item{total.variation}{logical; if \code{TRUE}, the standard errors
reported in the \code{Std. Error} column are based on the sum of
the likelihood variation and the MCMC variation. If \code{FALSE}
only the likelihood varuation is used. The \eqn{p}-values are
based on this source of variation.}
}
\value{
The function \code{\link{summary.ergm.tapered}} computes and
returns a list of summary statistics of the fitted
\code{\link{ergm}} model given in \code{object}. Note that for
backwards compatibility, it returns two coefficient tables:
\verb{$coefs} which does not contain the z-statistics and
\verb{$coefficeints} which does (and is therefore more similar to
those returned by \code{\link[=summary.lm]{summary.lm()}}).
}
\description{
\code{\link[base]{summary}} method for \code{\link{ergm.tapered}} fits.
}
\details{
\code{\link{summary.ergm.tapered}} tries to be smart about formatting the
coefficients, standard errors, etc.
}
\examples{

 data(florentine)

 x <- ergm(flomarriage ~ density)
 summary(x)

}
\seealso{
network, ergm, print.ergm.  The model fitting function
\code{\link{ergm}}, \code{\link{summary}}.

Function \code{\link{coef}} will extract the matrix of coefficients with
standard errors, t-statistics and p-values.
}
\keyword{models}
\keyword{regression}
