% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.Var.R
\name{Var-ergmTerm}
\alias{Var-ergmTerm}
\alias{InitErgmTerm.Var}
\title{Var}
\usage{
# binary: Var(formula=NULL, coef=NULL, m=NULL)
}
\arguments{
\item{formula}{a valid formula for a  standard ERGM}

\item{coef}{a numeric vector of coefficients giving the penalty coefficients \eqn{\beta} for the tapering of the terms.
If \code{NULL} is passed, the tapering coefficients are set to \code{1/(4*summary(formula))}, the default
in Fellows and Handcock (2017).
If a numeric vector is given, there are interpreted as the tapering coefficients of the terms in the
model, including the terms enclosed in \code{offset()}.
If a numeric scalar is given, it is interpreted as a multiplier of the default tapering coefficients , that is,}

\item{m}{numeric vector. If given, is
the value of the network statistic relative to which the model is
tapered. If omitted, it defaults to that of the model's LHS
network if \code{formula} is one-sided and that of the network on
the LHS of \code{formula} if it is two-sided.}
}
\description{
Adds the variance of the terms specified in \code{formula} to the model.
This is sometimes applied directly to control the variance of the standard
statistics, but more typically as part of a Tapered ERGM.
}
\references{
\itemize{
\item Fellows, I and Handcock, MS (2017). Removing Phase Transitions
from Gibbs Measures. \emph{Proceedings of Machine Learning
Research}, 54:289-297.
\item Blackburn, B and Handcock, MS (2022). Practical Network Modeling
via Tapered Exponential-family Random Graph Models. \emph{Journal
of Computational and Graphical Statistics}
\doi{10.1080/10618600.2022.2116444}.
}
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.
}
\concept{operator}
\concept{tapered}
