% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.ergm.tapered.R
\name{control.ergm.tapered}
\alias{control.ergm.tapered}
\title{Auxiliary for Controlling Tapered ERGM Fitting}
\usage{
control.ergm.tapered(
  MCMLE.metric = c("lognormal", "logtaylor", "Median.Likelihood", "EF.Likelihood",
    "naive", "Kpenalty"),
  MCMC.esteq.exclude.statistics = NULL,
  MCMLE.kurtosis.location = 3,
  MCMLE.kurtosis.scale = 0.3,
  MCMLE.kurtosis.penalty = 2,
  MCMLE.termination = c("precision", "confidence", "Hummel", "Hotelling", "none"),
  MCMLE.MCMC.precision = 0.005,
  loglik = control.logLik.ergm.tapered(),
  ...
)
}
\arguments{
\item{MCMLE.metric}{Method to calculate the loglikelihood approximation.}

\item{MCMC.esteq.exclude.statistics}{vector vector of names of statistics to exclude from the estimating equations.}

\item{MCMLE.kurtosis.location}{numeric The mean of the prior for kurtosis and/or the
target kurtosis value.}

\item{MCMLE.kurtosis.scale}{numeric The standard deviation of the prior for kurtosis.}

\item{MCMLE.kurtosis.penalty}{numeric: The size of the penalty for larger values of the tapering parameter on the log-likelihood.
See \code{\link[=control.ergm]{control.ergm()}} for the standard \code{\link[=ergm]{ergm()}} options. The tapering models add
\code{} and .}

\item{MCMLE.termination}{The criterion used for terminating MCMLE
estimation. The default for \code{ergm.tapered} is
\code{"precision"}: Terminate when the estimated loss in estimating precision
due to using MCMC standard errors is below the precision bound specified by
\code{MCMLE.MCMC.precision}, and the Hummel step length is 1 for two
consecutive iterations. See \code{MCMLE.MCMC.precision} for details. This
feature is in experimental status until we verify the coverage of the
standard errors. See the documentation for \code{\link{control.ergm}} for the other options}

\item{MCMLE.MCMC.precision}{\code{MCMLE.MCMC.precision} is a vector of upper bounds on the standard
errors induced by the MCMC algorithm, expressed as a percentage of the total
standard error. The MCMLE algorithm will terminate when the MCMC standard
errors are below the precision bound, and the Hummel step length is 1 for
two consecutive iterations. This is an experimental feature \code{\link[=ergm]{ergm()}}.
The default value in \code{\link[=ergm.tapered]{ergm.tapered()}} is 0.15, higher than in \code{\link[=ergm]{ergm()}}.}

\item{loglik}{list List of additional control arguments for the loglik. See \code{\link{control.logLik.ergm.tapered}}.}

\item{\dots}{Additional arguments, passed to other functions This argument
is helpful because it collects any control parameters that have been
deprecated; a warning message is printed in case of deprecated arguments.}

\item{MCMLE.effectiveSize.maxruns}{integer The maximum number of times the MCMLE.effectiveSize value will be doubled.}
}
\value{
A list with arguments as components. It includes the \code{\link[=control.ergm]{control.ergm()}} arguments.
}
\description{
Auxiliary function as user interface for fine-tuning 'ergm' fitting
for tapered models.
}
\details{
This function is only used within a call to the \code{\link[=ergm.tapered]{ergm.tapered()}} function.
See the \code{usage} section in \code{\link[=ergm]{ergm()}} for details.
}
\references{
\itemize{
\itemize{
\item Fellows, I. and M. S. Handcock (2017),
Removing Phase Transitions from Gibbs Measures. Volume 54 of
Proceedings of Machine Learning Research, Fort Lauderdale,
FL, USA, pp. 289–297. PMLR.
}

}
}
\seealso{
\code{\link[=control.ergm]{control.ergm()}}.
}
\keyword{models}
