% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.logLik.ergm.tapered.R
\name{control.logLik.ergm.tapered}
\alias{control.logLik.ergm.tapered}
\title{Auxiliary for Controlling Tapered logLik.ergm}
\usage{
control.logLik.ergm.tapered(
  MCMLE.metric = c("lognormal", "logtaylor", "Median.Likelihood", "EF.Likelihood",
    "naive", "Kpenalty"),
  MCMC.esteq.exclude.statistics = NULL,
  MCMLE.kurtosis.prior = TRUE,
  MCMLE.kurtosis.location = 3,
  MCMLE.kurtosis.scale = 0.3,
  MCMLE.kurtosis.equality = FALSE,
  ...
)
}
\arguments{
\item{MCMLE.metric}{Method to calculate the loglikelihood approximation.
See \code{\link[=control.ergm]{control.ergm()}} for the standard \code{\link[=ergm]{ergm()}} options. The tapering models add
\code{} and .}

\item{MCMLE.kurtosis.prior}{Logical: If TRUE, use a prior for the kurtosis.}

\item{MCMLE.kurtosis.location}{numeric The mean of the prior for kurtosis and/or the
taget kurtosis value.}

\item{MCMLE.kurtosis.scale}{numeric The standard deviation of the prior for kurtosis.}

\item{MCMLE.kurtosis.equality}{Logical: If TRUE, constrain the kurtosis in expectation.}

\item{\dots}{Additional arguments, passed to other functions This argument
is helpful because it collects any control parameters that have been
deprecated; a warning message is printed in case of deprecated arguments.}
}
\value{
A list with arguments as components. It includes the \code{\link[=control.ergm]{control.ergm()}} arguments.
}
\description{
Auxiliary function as user interface for fine-tuning logLik.ergm algorithm,
which approximates log likelihood values.
}
\details{
This function is only used within a call to the \code{\link{logLik.ergm}}
function.
}
\references{
\itemize{
\itemize{
\item Fellows, I. and M. S. Handcock (2017),
Removing Phase Transitions from Gibbs Measures. Volume 54 of
Proceedings of Machine Learning Research, Fort Lauderdale,
FL, USA, pp. 289–297. PMLR.
}

}
}
\seealso{
\code{\link[=control.ergm]{control.ergm()}}.
}
\keyword{models}
