% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.tapered.R
\name{ergm.tapered}
\alias{ergm.tapered}
\title{Fits a Tapered ERGM}
\usage{
ergm.tapered(
  formula,
  r = 2,
  beta = NULL,
  tau = NULL,
  tapering.centers = NULL,
  target.stats = NULL,
  family = "taper",
  taper.terms = "dependent",
  control = control.ergm(MCMLE.termination = "confidence"),
  ...
)
}
\arguments{
\item{formula}{An ergm formula to fit}

\item{r}{The scaling factor to use for the hueristic of setting beta equal to r standard deviations of the observed statistics}

\item{beta}{The tapering parameters, expressed as in Fellows and Handcock (2017). If not NULL, these override the hueristics (r).}

\item{tau}{The tapering parameters, expressed as natural parameters. If not NULL, these override the beta and the hueristics (r).}

\item{tapering.centers}{The centers of the tapering terms. If null, these are taken to be the mean value parameters.}

\item{target.stats}{{vector of "observed network statistics,"
if these statistics are for some reason different than the
actual statistics of the network on the left-hand side of
\code{formula}.
Equivalently, this vector is the mean-value parameter values for the
model.  If this is given, the algorithm finds the natural
parameter values corresponding to these mean-value parameters.
If \code{NULL}, the mean-value parameters used are the observed
statistics of the network in the formula.
}}

\item{family}{The type of tapering used. This should either be the \code{stereo} or \code{taper}, the
tapering model of Fellows and Handcock (2016).}

\item{taper.terms}{Specification of the tapering used. If the character variable "dependence" then all the dependent
terms are tapered. If the character variable "all" then all terms are tapered.
It can also be the RHS of a formula giving the terms to be tapered.}

\item{control}{An object of class control.ergm. Passed to the ergm function.}

\item{...}{Additional arguments to ergm.}
}
\value{
An object of class c('tapered.ergm','ergm') containing the fit model. In addition to all of the ergm items,
this object contains tapering.centers, tapering.coef and orig.formula. tapering.centers are the centers for the tapering term.
tapering.coef are the tapering coefficients = 1/ beta^2. orig.formula is the formula passed into ergm.tapered.
}
\description{
Fits a Tapered ERGM
}
\examples{
\dontrun{
data(sampson)
fit <- ergm.tapered(samplike ~ edges + triangles(), tau=0.1)
summary(fit)
}
}
\references{
Fellows, I and Handcock, MS (2017). Removing Phase Transitions from Gibbs Measures. Proceedings of Machine Learning Research, 54:289-297.
}
