% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.ergm.tapered.R
\name{gof.tapered.ergm}
\alias{gof.tapered.ergm}
\title{Fits a Tapered ERGM}
\usage{
\method{gof}{tapered.ergm}(object, ...)
}
\arguments{
\item{...}{Additional arguments to ergm.}

\item{formula}{An ergm formula to fit}

\item{r}{The scaling factor to use for the hueristic of setting beta equal to r the standard deviation of the observed statistics}

\item{beta}{The tapering parameters. If not null, these override the hueristics.}

\item{tapering.centers}{The centers of the tapering terms. If null, these are taken to be the mean value parameters.}

\item{control}{An object of class control.ergm. Passed to the ergm function.}
}
\description{
Fits a Tapered ERGM
}
\examples{
\dontrun{
data(sampson)
fit <- ergm.tapered(samplike ~ edges + triangles())
summary(fit)
}
}
\references{
Fellows, I and Handcock, MS (2017). Removing Phase Transitions from Gibbs Measures. Proceedings of Machine Learning Research, 54:289-297.
}
