% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.ergm.tapered.R,
%   R/summary.ergm.tapered.R
\name{print.summary.ergm.tapered}
\alias{print.summary.ergm.tapered}
\alias{summary.ergm.tapered}
\title{Summarizing Tapered and Kurtosis ERGM Model Fits}
\usage{
\method{print}{summary.ergm.tapered}(
  x,
  digits = max(3, getOption("digits") - 3),
  correlation = x$correlation,
  covariance = x$covariance,
  extended = x$extended,
  signif.stars = getOption("show.signif.stars"),
  eps.Pvalue = 1e-04,
  print.formula = FALSE,
  print.fitinfo = TRUE,
  print.coefmat = TRUE,
  print.message = TRUE,
  print.deviances = TRUE,
  print.drop = TRUE,
  print.offset = TRUE,
  print.call = TRUE,
  ...
)

\method{summary}{ergm.tapered}(
  object,
  ...,
  correlation = FALSE,
  covariance = FALSE,
  total.variation = TRUE,
  extended = FALSE
)
}
\arguments{
\item{x}{object of class \code{summary.ergm.tapered} returned by \code{\link[=summary.ergm.tapered]{summary.ergm.tapered()}}.}

\item{digits}{significant digits for coefficients}

\item{correlation}{logical; if \code{TRUE}, the correlation matrix
of the estimated parameters is returned and printed.}

\item{covariance}{logical; if \code{TRUE}, the covariance matrix of
the estimated parameters is returned and printed.}

\item{extended}{logical; if \code{TRUE}, the natural parameters of the tapered terms
are printed as an appended column of the summary table. Also printed are the bias
in the parameter values due to the tapering}

\item{signif.stars}{whether to print dots and stars to signify
statistical significance. See \code{\link[=print.summary.lm]{print.summary.lm()}}.}

\item{eps.Pvalue}{\eqn{p}-values below this level will be printed
as "<\code{eps.Pvalue}".}

\item{print.formula, print.fitinfo, print.coefmat, print.message, print.deviances, print.drop, print.offset, print.call}{which components of the fit summary to print.}

\item{\dots}{Arguments to \code{\link{logLik.ergm}}}

\item{object}{an object of class \code{"ergm.tapered"}, usually, a result
of a call to \code{\link{ergm.tapered}}.}

\item{total.variation}{logical; if \code{TRUE}, the standard errors
reported in the \code{Std. Error} column are based on the sum of
the likelihood variation and the MCMC variation. If \code{FALSE}
only the likelihood varuation is used. The \eqn{p}-values are
based on this source of variation.}
}
\value{
The function \code{\link{summary.ergm.tapered}} computes and
returns a list of summary statistics of the fitted
\code{\link{ergm}} model given in \code{object}. Note that for
backwards compatibility, it returns two coefficient tables:
\verb{$coefs} which does not contain the z-statistics and
\verb{$coefficients} which does (and is therefore more similar to
those returned by \code{\link[=summary.lm]{summary.lm()}}).
}
\description{
\code{\link[base]{summary}} method for \code{\link{ergm.tapered}} fits.
}
\details{
The default printout of the summary object contains the
call, number of iterations used, null and residual deviances, and
the values of AIC and BIC (and their MCMC standard errors, if
applicable). The coefficient table contains the following
columns:
\itemize{
\item \code{Estimate}, \verb{Std. Error} - parameter estimates and their standard errors
\item \verb{MCMC \%} - if \code{total.variation=TRUE} (default) the percentage of standard
error attributable to MCMC estimation process rounded to an integer. See
also \code{\link[=vcov.ergm]{vcov.ergm()}} and its \code{sources} argument.
\item \verb{z value}, \verb{Pr(>|z|)} - z-test and p-values
}

\code{\link{summary.ergm.tapered}} tries to be smart about formatting the
coefficients, standard errors, etc.
}
\examples{

data(sampson)
fit <- ergm.tapered(samplike ~ edges + triangles())
summary(fit)

}
\seealso{
network, ergm, print.ergm.  The model fitting function
\code{\link{ergm}}, \code{\link{summary}}.

Function \code{\link{coef}} will extract the matrix of coefficients with
standard errors, t-statistics and p-values.
}
\keyword{models}
\keyword{regression}
